+++ federation-integration-testsuite-js/CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG for `federation-integration-testsuite-js`
 
+## 2.4.6
+
 ## 2.4.5
 ### Patch Changes